chore(hog-charts): move generic utils + AnomalyPointsLayer to package level
Pure restructure, no behaviour change:

- Move AnomalyPointsLayer from charts/TimeSeriesLineChart/overlays/ up to
  lib/hog-charts/overlays/ alongside ReferenceLine and ValueLabels — it's
  a generic primitive, not TimeSeriesLineChart-specific.
- Move dates, y-formatters, goal-lines, and use-axis-formatters from
  charts/TimeSeriesLineChart/utils/ to lib/hog-charts/utils/ — they're
  cross-cutting helpers used outside TimeSeriesLineChart.
- Extract applyComparisonDimming + dimHex into utils/comparison-dimming.ts
  (with its own test file). derived-series.ts keeps the CI / moving-avg /
  trend-line builders and imports dimHex from the new module.
- Update barrel re-exports and the one deep-path import in
  trendsAxisFormat.ts to consume YFormatterConfig from the barrel.

All renames use git mv so blame is preserved.
