+++ frontend/src/lib/hog-charts/utils/use-axis-formatters.ts
